webCrawler/requirements.txt
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
cssselect==1.1.0
idna==3.3
lxml==4.6.4
munch==2.5.0
Pillow==8.4.0
pyquery==1.4.3
pytesseract==0.3.8
requests==2.26.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.7
